Fix constraint drift detection
The previous configuration of triggers at the edge of detection was not sensitive enough to detect
all cases of constraint drift. Instead, we now place the constraint detector transmitter at a known
proximity from the receiver, on the same direction vector as all other values. This helps ensure that
we can detect when the constraints drift off their target position accurately.

Closes: #1